+++ example/particle_leapfrog/particle_leapfrog.hpp
@@ -1,6 +1,6 @@
 //========================================================================================
 // Parthenon performance portable AMR framework
-// Copyright(C) 2021 The Parthenon collaboration
+// Copyright(C) 2021-2026 The Parthenon collaboration
 // Licensed under the 3-clause BSD License, see LICENSE file for details
 //========================================================================================
 // (C) (or copyright) 2020-2021. Triad National Security, LLC. All rights reserved.
@@ -44,6 +44,17 @@ class ParticleDriver : public EvolutionDriver {
 
 void ProblemGenerator(MeshBlock *pmb, ParameterInput *pin);
 Packages_t ProcessPackages(std::unique_ptr<ParameterInput> &pin);
+
+//----------------------------------------------------------------------------------------
+//! \class ParticleUserOutput
+//  \brief derived OutputType class for User enrolled outputs
+
+class ParticleUserOutput : public OutputType {
+ public:
+  explicit ParticleUserOutput() : OutputType({}) {}
+  void WriteOutputFile(Mesh *pm, ParameterInput *pin, SimTime *tm,
+                       const SignalHandler::OutputSignal signal) override;
+};
 
 namespace Particles {