+++ src/interface/outputs.hpp
@@ -1,6 +1,6 @@
 //========================================================================================
 // Parthenon performance portable AMR framework
-// Copyright(C) 2020-2025 The Parthenon collaboration
+// Copyright(C) 2020-2026 The Parthenon collaboration
 // Licensed under the 3-clause BSD License, see LICENSE file for details
 //========================================================================================
 // Athena++ astrophysical MHD code
@@ -18,16 +18,23 @@
 // license in this material to reproduce, prepare derivative works, distribute copies to
 // the public, perform publicly and display publicly, and to permit others to do so.
 //========================================================================================
-
-#ifndef OUTPUTS_OUTPUT_PARAMETERS_HPP_
-#define OUTPUTS_OUTPUT_PARAMETERS_HPP_
+#ifndef INTERFACE_OUTPUTS_HPP_
+#define INTERFACE_OUTPUTS_HPP_
+//! \file outputs.hpp
+//  \brief provides base classes to handle ALL types of data output
 
 #include <map>
 #include <set>
 #include <string>
 #include <vector>
 
+#include "basic_types.hpp"
+
 namespace parthenon {
+
+// forward declarations
+class Mesh;
+class ParameterInput;
 
 // JMM: I designed this for HDF5 but in pinciple this switching could
 // also work for other output types... Any output type that is capable
@@ -74,6 +81,41 @@ struct OutputParameters {
   int file_number = 0;
 };
 
-} // namespace parthenon
+//----------------------------------------------------------------------------------------
+//  \brief abstract base class for different output types (modes/formats). Each OutputType
+//  is designed to be a node in a singly linked list created & stored in the Outputs class
 
-#endif // OUTPUTS_OUTPUT_PARAMETERS_HPP_
+class OutputType {
+ public:
+  // mark single parameter constructors as "explicit" to prevent them from acting as
+  // implicit conversion functions: for f(OutputType arg), prevent f(anOutputParameters)
+  explicit OutputType(OutputParameters oparams);
+
+  // rule of five:
+  virtual ~OutputType() = default;
+  // copied)
+  OutputType(const OutputType &copy_other) = default;
+  OutputType &operator=(const OutputType &copy_other) = default;
+  // move constructor and assignment operator
+  OutputType(OutputType &&) = default;
+  OutputType &operator=(OutputType &&) = default;
+
+  // data
+  OutputParameters output_params; // control data read from <output> block
+
+  // following pure virtual function must be implemented in all derived classes
+  virtual void WriteOutputFile(Mesh *pm, ParameterInput *pin, SimTime *tm,
+                               const SignalHandler::OutputSignal signal) = 0;
+  virtual void WriteContainer(SimTime &tm, Mesh *pm, ParameterInput *pin, bool flag) {
+    return;
+  }
+
+ protected:
+  int num_vars_; // number of variables in output
+
+  // Update book-keeping such as next output time to next output
+  void UpdateNextOutput_(Mesh *pm, SimTime *tm);
+};
+
+} // namespace parthenon
+#endif // INTERFACE_OUTPUTS_HPP_